Set the binary type when deserializing and ELF header from GTIRB
When serializing a GTIRB file from ELF we set the binary type from
the ELF header.  However, there is no corresponding conversion
to set the binary type in the ELF header when deserializing from GTIRB.

Create a static method getBinaryTypeFlag to parse the type flag using
the same rules as when serializing, and set the e_type flag using
this method.